# refactor: build aggregates in a column-oriented fashion 