This PR changes how aggregates and grouped aggregates are build in the Read Buffer.

## Reason for Change

The reason for this change is twofold:

1) Primarily it will simplify adding support for `FIRST` and `LAST` to the Read Buffer as supported aggregations. Possibly also `MEAN` if we decide to push that down too.
2) It improves the memory footprint and query performance of `read_aggregate` operations (any query that does an aggregate or has a group by in it).

## Background

Previously the Read Buffer took a simple approach of building grouped aggregates in a row-wise way. Effectively the read buffer would materialise all of the columns that were being aggregated and also materialise compressed versions of any columns being grouped on.

Then it would iterate the grouping columns, building group keys in a hashmap. As it did this it would store aggregations as a value to each key in the map. Whenever the group key already existed for a new row being processed, the aggregates would be updated with the values from the input aggregation columns at the same  row offset.

This approach works fine but it means a few things:

1) there is excessive branching caused by having to match on many enums each time you want to update an aggregate (you have to figure out what the data type is and what the aggregation operation is).
2) you have to take these row-wise results and pivot them into column-wise results when they leave the read buffer via record batches.
3) First/Last would be annoying to implement because you need to track a tuple for each row (the value and timestamp) where as the existing implementation only supported aggregates having a single value per row.

## Change

The Read Buffer now allocates vectors for each output group column and each output group aggregate. Then instead of storing group keys and the aggregates in a hashmap we instead store group keys and the _ordinal offset into output columns_. As we're processing input columns to execute the query we use each group key's offset to update the aggregates in the pre-allocated output columns. 

Once the operation is complete we then no longer need to do any pivoting and we have the data read in column form to convert to a record batch.

Further, this approach means for first/last we can store two output vectors (one for the candidate values and one for their timestamps).

There is quite a lot of code in this PR but the net change won't be too bad as I will be able to delete a bunch of code that's no longer needed on subsequent PRs.

## Results

One of the nice side-effects of this reworking is that there is an improvement in performance, both in terms of memory and CPU.

I tested the changes on my tracing dataset. This is a 10 million row dataset that has a cardinality of ~10M.

I execution of the equivalent of these two queries:

```
A: SELECT SUM("duration") FROM "tracing" GROUP BY "trace_id";

B: SELECT SUM("duration") FROM "tracing" GROUP BY "trace_id", "env";
```

| Query Type  | before: mem allocated | PR: mem allocated  | before: time | PR: time  |
|---|---|---|---|---|
| A  | 150MB  | 35MB  | 480 ms  | 420 ms (12.5% faster)  |
|  B | 95MB  | 52MB  | 590 ms  | 479 ms (19% faster) |

The approach in this PR is significantly better in terms of memory and CPU.

## Next Steps 

- [ ] Remove some dead code.
- [ ] Implement First/Last aggregates
- [ ] Implement windowed aggregates
- [ ] Refactor `row_group.rs` to move code into modules as the file is getting way too big now.