@e-dard 2021-03-22
@alamb wrote:

> It seems like the improvement in memory usage can be explained by avoiding two copies of the output (the copy in the hash table and the copy in the output vector). Do you think that avoiding the output pivot is also the source of the CPU improvement? Or is it the more sophisticated handling of merging group keys?

Other factors include: fewer allocations because the aggregates are kept in a single column-wise vector rather than many row-wise vectors, and fewer matches due to coalescing the datatype and aggregation type via `AggregateVec`. This mean less branching, which very likely helps with throughput.

I think I got all or most of the comments so will merge on 🍏